카페인 계산 수정(몸무게 추가)

+++ src/main/java/com/latte/drink/repository/DrinkMapper.java
@@ -29,8 +29,8 @@ public interface DrinkMapper {
 
     DateStatusResponse findSumCaffeineByDate(@Param("mbrNo") int mbrNo,
                                              @Param("localDateTime") LocalDateTime localDateTime,
-                                             @Param("minNormal") int minNormal,
-                                             @Param("maxNormal") int maxNormal);
+                                             @Param("minNormal") double minNormal,
+                                             @Param("maxNormal") double maxNormal);
 
     List<DrinkMenuResponse> findMenuByDate(@Param("mbrNo") int mbrNo,
                                            @Param("localDateTime") LocalDateTime localDateTime);
